backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
anthropic==0.40.0
python-dotenv==1.0.1
python-docx==1.1.2
pdfplumber==0.11.0
python-pptx==1.0.2
openpyxl==3.1.2
beautifulsoup4==4.12.3
Pillow==10.4.0
